Refactor DescribeCommandTest to use anonymous class for "testCommandDescribesCustomFixer"
