--- tests/Fixtures/DescribeCommand/.php-cs-fixer.fixture.php
@@ -1,9 +1,0 @@
-<?php
-
-use PhpCsFixer\Config;
-use PhpCsFixer\Finder;
-use PhpCsFixer\Tests\Fixtures\DescribeCommand\DescribeFixtureFixer;
-
-return (new Config())
-    ->setFinder((new Finder())->in(__DIR__))
-    ->registerCustomFixers([new DescribeFixtureFixer()]);
